requirements.txt
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
automaton==3.2.0
autopage==0.5.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
cliff==4.10.0
cmd2==2.7.0
comm==0.2.3
cryptography==45.0.6
debtcollector==3.0.0
debugpy==1.8.16
decorator==5.2.1
dnspython==2.7.0
dogpile.cache==1.4.0
eventlet==0.40.2
executing==2.2.0
faiss-cpu==1.12.0
fastapi==0.116.1
fasteners==0.20
filetype==1.2.0
futurist==3.2.0
gnureadline==8.2.13
google-ai-generativelanguage==0.6.18
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-genai==1.30.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.4.0
ipython_pygments_lexers==1.1.1
iso8601==2.1.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.8.1
keystoneauth1==5.11.1
keystonemiddleware==10.11.0
kombu==5.5.4
langchain==0.3.27
langchain-core==0.3.74
langchain-google-genai==2.1.9
langchain-text-splitters==0.3.9
langsmith==0.4.15
legacy-cgi==2.6.3
magnum==20.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.1
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.5
numpy==2.3.2
ollama==0.5.3
openstacksdk==4.7.0
orjson==3.11.2
os-client-config==2.3.0
os-service-types==1.8.0
osc-lib==4.1.0
oslo.cache==3.11.0
oslo.concurrency==7.1.0
oslo.config==10.0.0
oslo.context==6.0.0
oslo.db==17.4.0
oslo.i18n==6.5.1
oslo.log==7.2.0
oslo.messaging==17.1.0
oslo.metrics==0.12.0
oslo.middleware==6.6.0
oslo.policy==4.5.1
oslo.reports==3.5.1
oslo.serialization==5.7.0
oslo.service==4.2.2
oslo.upgradecheck==2.5.0
oslo.utils==9.0.0
oslo.versionedobjects==3.7.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
Paste==3.10.1
PasteDeploy==3.1.0
pbr==7.0.0
pecan==1.7.0
pexpect==4.9.0
platformdirs==4.3.8
prettytable==3.16.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycadf==4.0.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
Pygments==2.19.2
PyJWT==2.10.1
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyperclip==1.9.0
python-barbicanclient==7.1.0
python-cinderclient==9.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-glanceclient==4.9.0
python-heatclient==4.3.0
python-keystoneclient==5.6.0
python-neutronclient==11.6.0
python-novaclient==18.10.0
python-octaviaclient==3.11.1
python-openstackclient==8.1.0
python-swiftclient==4.8.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
referencing==0.36.2
repoze.lru==0.7
requests==2.32.4
requests-toolbelt==1.0.0
requestsexceptions==1.4.0
rfc3986==2.0.0
rich==14.1.0
rich-argparse==1.7.1
Routes==2.5.1
rpds-py==0.27.0
rsa==4.9.1
setuptools==80.9.0
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
stack-data==0.6.3
starlette==0.47.2
statsd==4.0.1
stevedore==5.4.1
taskflow==6.0.1
tenacity==9.1.2
testresources==2.0.2
testscenarios==0.5.0
testtools==2.7.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
vine==5.1.0
warlock==2.1.0
wcwidth==0.2.13
WebOb==1.8.9
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
WSME==0.12.1
yappi==1.6.10
zipp==3.23.0
zstandard==0.24.0
